I give up
reasons:
i have no idea what this guy did to make the game so bad when upgraded.
the fade in / fade out animations have been structured in such a bad
way.
the ui has been changed in a way that would require a redo because of
godot 3 to 4 changes
building this project from scratch would be faster than trying to
upgrade it.
end of rant